+++ src/main/java/subway/domain/edge/Edges.java
@@ -1,4 +1,6 @@
-package subway.domain;
+package subway.domain.edge;
+
+import subway.domain.station.Station;
 
 import java.util.ArrayList;
 import java.util.HashSet;
