+++ src/main/java/subway/domain/line/dto/LineCreateRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.domain.line.dto;
 
 public class LineCreateRequest {
     private final String lineName;
