+++ src/main/java/subway/service/LineService.java
@@ -5,13 +5,13 @@ import org.springframework.transaction.annotation.Transactional;
 import subway.dao.EdgeDao;
 import subway.dao.LineDao;
 import subway.dao.StationDao;
-import subway.domain.Edge;
-import subway.domain.Edges;
-import subway.domain.Line;
-import subway.domain.Lines;
-import subway.domain.Station;
-import subway.dto.AddStationToLineRequest;
-import subway.dto.LineCreateRequest;
+import subway.domain.edge.Edge;
+import subway.domain.edge.Edges;
+import subway.domain.line.Line;
+import subway.domain.line.Lines;
+import subway.domain.line.dto.AddStationToLineRequest;
+import subway.domain.line.dto.LineCreateRequest;
+import subway.domain.station.Station;
 
 import java.util.ArrayList;
 import java.util.HashMap;
