+++ src/main/java/subway/ui/StationController.java
@@ -1,15 +1,16 @@
 package subway.ui;
 
-import java.net.URI;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import subway.domain.Station;
-import subway.dto.StationCreateRequest;
-import subway.dto.StationCreateResponse;
+import subway.domain.station.Station;
+import subway.domain.station.dto.StationCreateRequest;
+import subway.domain.station.dto.StationCreateResponse;
 import subway.service.StationService;
+
+import java.net.URI;
 
 @RestController
 @RequestMapping("/stations")
