+++ src/test/java/subway/SubwayApplicationTests.java
@@ -9,5 +9,4 @@ class SubwayApplicationTests {
     @Test
     void contextLoads() {
     }
-
 }
